server/requirements.txt
fastapi[standard]>=0.115.14
logfire[fastapi]>=3.22.1
logfire[sqlalchemy]
pydantic-evals[logfire]
pydantic_ai
pydantic
gradio>=5.35.0
nest_asyncio
llama-index>=0.12.47
llama-index-core>=0.12.47
llama-index-readers-file
llama-index-readers-qdrant
llama-index-embeddings-ollama
llama-index-embeddings-openai
llama-index-vector-stores-qdrant
llama-index-llms-lmstudio
llama-index-llms-openai
llama-index-embeddings-openai-like
opentelemetry-instrumentation-llamaindex
spacy
sqlalchemy
psycopg[binary,pool]
ragchecker>=0.1.9